Add game for moving things with the keyboard
**This Commit**
Adds a new game where the player moves things around with their
keyboard. It has meshes in it. That's fancy. Needs lots of work though
I'm sure. We'll get some friends to play it and some nerds to look at it
and hopefully learn a thing about organizing some game code!

**Todo**
Come up with a cool polygon overlay algorithm so we can:

1. use triangle based meshes instead of line meshes
2. do something less "lucky" then just comparing translation/rotation (I
   still have no idea how the translation one Just Works)