Rebind right-hand player keys
**This Commit**
Uses I, J, K, L, U, and O for the right hand player movement instead of
the arrow keys.

**Why?**
Well it's slightly more vim-like so that's cool but the real reason is
that it's more symmetric and works on keyboards who might not have arrow
keys near their Shift + ? keys.

Co-authored-by: Brian Smith <smithbrian387@gmail.com>