Store mesh points on target and players (#4)
**Why?**
So we don't have to get the points out of the mesh attributes every
frame. This is slightly cleaner though I'm not sure if it's ideal. It at
least cleans up some system queries/parameters.

**Notes**
These entities taking up more space on the stack _might_ be a problem
but the target moves every frame and it's very likely that the player is
always moving the players so it seems that we'll need to get this off
the mesh every frame and hopefully it's not wasteful.